+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 * Add `EveOnline::ESI::Alliances`
 * Add `EveOnline::ESI::AllianceCorporations`
+* Add `EveOnline::ESI::Models::AllianceShort`
+* Add `EveOnline::ESI::AlliancesNames`
 
 **v0.14.0**